+++ src/main/java/project/flipnote/auth/controller/AuthController.java
@@ -18,16 +18,16 @@ import org.springframework.web.bind.annotation.RestController;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import project.flipnote.auth.controller.docs.AuthControllerDocs;
-import project.flipnote.auth.model.ChangePasswordRequest;
-import project.flipnote.auth.model.EmailVerificationRequest;
-import project.flipnote.auth.model.EmailVerifyRequest;
-import project.flipnote.auth.model.PasswordResetCreateRequest;
-import project.flipnote.auth.model.PasswordResetRequest;
-import project.flipnote.auth.model.TokenPair;
-import project.flipnote.auth.model.UserLoginRequest;
-import project.flipnote.auth.model.UserLoginResponse;
-import project.flipnote.auth.model.UserRegisterRequest;
-import project.flipnote.auth.model.UserRegisterResponse;
+import project.flipnote.auth.model.request.ChangePasswordRequest;
+import project.flipnote.auth.model.request.EmailVerificationRequest;
+import project.flipnote.auth.model.request.EmailVerifyRequest;
+import project.flipnote.auth.model.request.PasswordResetCreateRequest;
+import project.flipnote.auth.model.request.PasswordResetRequest;
+import project.flipnote.auth.model.vo.TokenPair;
+import project.flipnote.auth.model.request.UserLoginRequest;
+import project.flipnote.auth.model.response.UserLoginResponse;
+import project.flipnote.auth.model.request.UserRegisterRequest;
+import project.flipnote.auth.model.response.UserRegisterResponse;
 import project.flipnote.auth.service.AuthService;
 import project.flipnote.common.security.dto.AuthPrinciple;
 import project.flipnote.common.security.jwt.JwtConstants;
