+++ src/main/java/project/flipnote/auth/controller/OAuthController.java
@@ -21,8 +21,8 @@ import lombok.extern.slf4j.Slf4j;
 import project.flipnote.auth.constants.OAuthConstants;
 import project.flipnote.auth.controller.docs.OAuthControllerDocs;
 import project.flipnote.auth.exception.AuthErrorCode;
-import project.flipnote.auth.model.AuthorizationRedirect;
-import project.flipnote.auth.model.TokenPair;
+import project.flipnote.auth.model.vo.AuthorizationRedirect;
+import project.flipnote.auth.model.vo.TokenPair;
 import project.flipnote.auth.service.OAuthService;
 import project.flipnote.common.config.ClientProperties;
 import project.flipnote.common.exception.BizException;
