+++ src/main/java/project/flipnote/auth/listener/EmailVerificationEventListener.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import project.flipnote.auth.constants.VerificationConstants;
-import project.flipnote.auth.event.EmailVerificationSendEvent;
+import project.flipnote.auth.model.event.EmailVerificationSendEvent;
 import project.flipnote.common.exception.EmailSendException;
 import project.flipnote.infra.email.EmailService;
