+++ src/main/java/project/flipnote/auth/listener/PasswordResetCreateEventListener.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import project.flipnote.auth.constants.PasswordResetConstants;
-import project.flipnote.auth.event.PasswordResetCreateEvent;
+import project.flipnote.auth.model.event.PasswordResetCreateEvent;
 import project.flipnote.common.exception.EmailSendException;
 import project.flipnote.infra.email.EmailService;
