+++ src/main/java/project/flipnote/auth/model/event/EmailVerificationSendEvent.java
@@ -1,4 +1,4 @@
-package project.flipnote.auth.event;
+package project.flipnote.auth.model.event;
 
 public record EmailVerificationSendEvent(
 	String to,
