+++ src/main/java/project/flipnote/auth/model/request/PasswordResetRequest.java
@@ -1,4 +1,4 @@
-package project.flipnote.auth.model;
+package project.flipnote.auth.model.request;
 
 import jakarta.validation.constraints.NotBlank;
 import project.flipnote.common.validation.annotation.ValidPassword;
