+++ src/main/java/project/flipnote/auth/model/vo/AuthorizationRedirect.java
@@ -1,4 +1,4 @@
-package project.flipnote.auth.model;
+package project.flipnote.auth.model.vo;
 
 import org.springframework.http.ResponseCookie;
