+++ src/main/java/project/flipnote/auth/service/AuthService.java
@@ -13,17 +13,17 @@ import project.flipnote.auth.constants.VerificationConstants;
 import project.flipnote.auth.entity.AccountStatus;
 import project.flipnote.auth.entity.OAuthLink;
 import project.flipnote.auth.entity.UserAuth;
-import project.flipnote.auth.event.EmailVerificationSendEvent;
-import project.flipnote.auth.event.PasswordResetCreateEvent;
-import project.flipnote.auth.model.ChangePasswordRequest;
-import project.flipnote.auth.model.EmailVerificationRequest;
-import project.flipnote.auth.model.EmailVerifyRequest;
-import project.flipnote.auth.model.PasswordResetCreateRequest;
-import project.flipnote.auth.model.PasswordResetRequest;
-import project.flipnote.auth.model.TokenPair;
-import project.flipnote.auth.model.UserLoginRequest;
-import project.flipnote.auth.model.UserRegisterRequest;
-import project.flipnote.auth.model.UserRegisterResponse;
+import project.flipnote.auth.model.event.EmailVerificationSendEvent;
+import project.flipnote.auth.model.event.PasswordResetCreateEvent;
+import project.flipnote.auth.model.request.ChangePasswordRequest;
+import project.flipnote.auth.model.request.EmailVerificationRequest;
+import project.flipnote.auth.model.request.EmailVerifyRequest;
+import project.flipnote.auth.model.request.PasswordResetCreateRequest;
+import project.flipnote.auth.model.request.PasswordResetRequest;
+import project.flipnote.auth.model.vo.TokenPair;
+import project.flipnote.auth.model.request.UserLoginRequest;
+import project.flipnote.auth.model.request.UserRegisterRequest;
+import project.flipnote.auth.model.response.UserRegisterResponse;
 import project.flipnote.auth.repository.EmailVerificationRedisRepository;
 import project.flipnote.auth.repository.OAuthLinkRepository;
 import project.flipnote.auth.repository.PasswordResetRedisRepository;
