+++ src/main/java/project/flipnote/auth/service/OAuthService.java
@@ -14,8 +14,8 @@ import lombok.extern.slf4j.Slf4j;
 import project.flipnote.auth.constants.OAuthConstants;
 import project.flipnote.auth.entity.OAuthLink;
 import project.flipnote.auth.exception.AuthErrorCode;
-import project.flipnote.auth.model.AuthorizationRedirect;
-import project.flipnote.auth.model.TokenPair;
+import project.flipnote.auth.model.vo.AuthorizationRedirect;
+import project.flipnote.auth.model.vo.TokenPair;
 import project.flipnote.auth.repository.OAuthLinkRepository;
 import project.flipnote.auth.repository.SocialLinkTokenRedisRepository;
 import project.flipnote.auth.repository.UserAuthRepository;
